+++ client/src/pages/VenueDetails.jsx
@@ -69,7 +69,7 @@ const VenueDetails = ({ match }) => {
                         <FaStar className="text-yellow-500 ml-4" /> {venue.rating} ({venue.reviews?.length || 0})
                     </div>
                 </div>
-                <Link to={"/venue"}>
+                <Link to={"/booking"}>
                     <button className="bg-green-500 text-white px-6 py-2 rounded hover:bg-green-600 mt-4 md:mt-0">
                         Book This Venue
                     </button>
